requirements.txt
streamlit
#for our ui
google-generativeai
#for accessing prompt
python-dotenv
#to run api keys from .env